remove: delete
- 필요 없다고 판단
- display와 합병
